+++ src/main/java/com/example/demo/service/UserService.java
@@ -5,5 +5,5 @@ import com.example.demo.entity.dto.UserDTO;
 public interface UserService {
     void register(UserDTO userDTO) throws Exception;
     UserDTO login(String username, String password) throws Exception;
-    UserDTO findById(Integer id) throws Exception;
+    UserDTO findById(Long id) throws Exception;
 }
